update code module
