+++ apps/web/client/src/components/store/editor/engine/code/index.ts
@@ -4,33 +4,17 @@ import { sendAnalytics } from '@/utils/analytics';
 import type {
     Action,
     CodeDiff, CodeDiffRequest,
-    CodeGroup,
-    CodeInsertImage,
-    CodeMove,
-    CodeRemove,
-    CodeRemoveImage,
-    CodeUngroup,
-    EditTextAction,
-    GroupElementsAction,
-    InsertElementAction,
-    InsertImageAction,
-    MoveElementAction,
-    RemoveElementAction,
-    RemoveImageAction,
-    UngroupElementsAction,
-    UpdateStyleAction,
-    WriteCodeAction,
+    WriteCodeAction
 } from '@onlook/models';
-import { CodeActionType } from '@onlook/models';
 import { assertNever } from '@onlook/utility';
+import { debounce } from 'lodash';
 import { makeAutoObservable } from 'mobx';
-import { getOrCreateCodeDiffRequest } from './helpers';
-import { getInsertedElement } from './insert';
-import { addTailwindToRequest } from './tailwind';
+import { getEditTextRequests, getGroupRequests, getInsertImageRequests, getInsertRequests, getMoveRequests, getRemoveImageRequests, getRemoveRequests, getStyleRequests, getUngroupRequests } from './requests';
 
 export class CodeManager {
     isExecuting = false;
     private writeQueue: Action[] = [];
+    private pendingRequests: CodeDiffRequest[] = [];
 
     constructor(
         private editorEngine: EditorEngine,
@@ -40,264 +24,73 @@ export class CodeManager {
     }
 
     async write(action: Action) {
-        // TODO: These can all be processed at once at the getCodeDiffRequests level
         this.writeQueue.push(action);
         if (!this.isExecuting) {
-            await this.processWriteQueue();
+            this.isExecuting = true;
+            this.debouncedProcessQueue();
         }
     }
 
-    private async processWriteQueue() {
-        this.isExecuting = true;
-        if (this.writeQueue.length > 0) {
-            const action = this.writeQueue.shift();
-            if (action) {
-                await this.executeWrite(action);
+    private processQueue = async () => {
+        try {
+            // Process all pending actions and collect the requests
+            while (this.writeQueue.length > 0) {
+                const action = this.writeQueue.shift();
+                if (action) {
+                    const requests = await this.collectRequests(action);
+                    if (Array.isArray(requests)) {
+                        this.pendingRequests.push(...requests);
+                    }
+                }
             }
-        }
-        setTimeout(() => {
+
+            // Write all collected requests at once
+            if (this.pendingRequests.length > 0) {
+                await this.writeRequest(this.pendingRequests);
+                this.pendingRequests = [];
+                sendAnalytics('write code');
+            }
+        } finally {
             this.isExecuting = false;
+            // Check if more actions were added while processing
             if (this.writeQueue.length > 0) {
-                this.processWriteQueue();
+                this.isExecuting = true;
+                this.debouncedProcessQueue();
             }
-        }, 300);
-    }
+        }
+    };
 
-    private async executeWrite(action: Action) {
+    private debouncedProcessQueue = debounce(this.processQueue, 300);
+
+    private async collectRequests(action: Action): Promise<CodeDiffRequest[]> {
         switch (action.type) {
             case 'update-style':
-                await this.writeStyle(action);
-                break;
+                return await getStyleRequests(action);
             case 'insert-element':
-                await this.writeInsert(action);
-                break;
+                return await getInsertRequests(action);
             case 'move-element':
-                await this.writeMove(action);
-                break;
+                return await getMoveRequests(action);
             case 'remove-element':
-                await this.writeRemove(action);
-                break;
+                return await getRemoveRequests(action);
             case 'edit-text':
-                await this.writeEditText(action);
-                break;
+                return await getEditTextRequests(action);
             case 'group-elements':
-                await this.writeGroup(action);
-                break;
+                return await getGroupRequests(action);
             case 'ungroup-elements':
-                await this.writeUngroup(action);
-                break;
-            case 'write-code':
-                await this.writeCode(action);
-                break;
+                return await getUngroupRequests(action);
             case 'insert-image':
-                await this.writeInsertImage(action);
-                break;
+                return getInsertImageRequests(action);
             case 'remove-image':
-                await this.writeRemoveImage(action);
-                break;
+                return getRemoveImageRequests(action);
+            case 'write-code':
+                return [];
             default:
                 assertNever(action);
         }
-        sendAnalytics('write code');
     }
 
-    async writeStyle({ targets }: UpdateStyleAction) {
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-
-        for (const target of targets) {
-            if (!target.oid) {
-                console.error('No oid found for style change');
-                continue;
-            }
-
-            const request = await getOrCreateCodeDiffRequest(target.oid, oidToCodeChange);
-            addTailwindToRequest(request, target.change.updated);
-        }
-
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    async writeInsert({ location, element, pasteParams, codeBlock }: InsertElementAction) {
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-        const insertedEl = getInsertedElement(element, location, pasteParams, codeBlock);
-
-        if (!insertedEl.location.targetOid) {
-            console.error('No oid found for inserted element');
-            return;
-        }
-
-        const request = await getOrCreateCodeDiffRequest(
-            insertedEl.location.targetOid,
-            oidToCodeChange,
-        );
-        request.structureChanges.push(insertedEl);
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    private async writeRemove({ element, codeBlock }: RemoveElementAction) {
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-        const removedEl: CodeRemove = {
-            oid: element.oid,
-            type: CodeActionType.REMOVE,
-            codeBlock,
-        };
-
-        const request = await getOrCreateCodeDiffRequest(removedEl.oid, oidToCodeChange);
-        request.structureChanges.push(removedEl);
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    private async writeEditText({ targets, newContent }: EditTextAction) {
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-
-        for (const target of targets) {
-            if (!target.oid) {
-                console.error('No oid found for text edit');
-                continue;
-            }
-            const request = await getOrCreateCodeDiffRequest(target.oid, oidToCodeChange);
-            request.textContent = newContent;
-        }
-
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    private async writeMove({ targets, location }: MoveElementAction) {
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-
-        for (const target of targets) {
-            if (!target.oid) {
-                console.error('No oid found for move');
-                continue;
-            }
-
-            if (!location.targetOid) {
-                console.error('No target oid found for moved element');
-                continue;
-            }
-
-            const movedEl: CodeMove = {
-                oid: target.oid,
-                type: CodeActionType.MOVE,
-                location,
-            };
-
-            const request = await getOrCreateCodeDiffRequest(location.targetOid, oidToCodeChange);
-            request.structureChanges.push(movedEl);
-        }
-
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    private async writeGroup(action: GroupElementsAction) {
-        if (!action.parent.oid) {
-            console.error('No parent oid found for group');
-            return;
-        }
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-        const groupEl: CodeGroup = {
-            type: CodeActionType.GROUP,
-            oid: action.parent.oid,
-            container: action.container,
-            children: action.children,
-        };
-
-        const request = await getOrCreateCodeDiffRequest(groupEl.oid, oidToCodeChange);
-        request.structureChanges.push(groupEl);
-
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    private async writeUngroup(action: UngroupElementsAction) {
-        if (!action.parent.oid) {
-            console.error('No parent oid found for ungroup');
-            return;
-        }
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-        const ungroupEl: CodeUngroup = {
-            type: CodeActionType.UNGROUP,
-            oid: action.parent.oid,
-            container: action.container,
-            children: action.children,
-        };
-
-        const request = await getOrCreateCodeDiffRequest(ungroupEl.oid, oidToCodeChange);
-        request.structureChanges.push(ungroupEl);
-
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    private async writeCode(action: WriteCodeAction) {
-        const res = await this.editorEngine.sandbox?.writeCodeDiffs(action.diffs);
-        if (!res) {
-            console.error('Failed to write code');
-            return false;
-        }
-        return true;
-    }
-
-    private async writeInsertImage(action: InsertImageAction) {
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-        const projectFolder = this.projectsManager.project?.folderPath;
-
-        if (!projectFolder) {
-            console.error('Failed to write image, projectFolder not found');
-            return;
-        }
-
-        const insertImage: CodeInsertImage = {
-            ...action,
-            folderPath: projectFolder,
-            type: CodeActionType.INSERT_IMAGE,
-        };
-
-        for (const target of action.targets) {
-            if (!target.oid) {
-                console.error('No oid found for inserted image');
-                continue;
-            }
-            const request = await getOrCreateCodeDiffRequest(target.oid, oidToCodeChange);
-            request.structureChanges.push(insertImage);
-        }
-
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    private async writeRemoveImage(action: RemoveImageAction) {
-        const oidToCodeChange = new Map<string, CodeDiffRequest>();
-        const removeImage: CodeRemoveImage = {
-            ...action,
-            type: CodeActionType.REMOVE_IMAGE,
-        };
-
-        for (const target of action.targets) {
-            if (!target.oid) {
-                console.error('No oid found for removed image');
-                continue;
-            }
-            const request = await getOrCreateCodeDiffRequest(target.oid, oidToCodeChange);
-            request.structureChanges.push(removeImage);
-        }
-
-        await this.writeRequest(Array.from(oidToCodeChange.values()));
-    }
-
-    async writeRequest(requests: CodeDiffRequest[], useHistory = false) {
-        let codeDiffs: CodeDiff[];
-        if (useHistory) {
-            codeDiffs = await this.editorEngine.sandbox?.getCodeDiffs(requests);
-            this.runCodeDiffs(codeDiffs);
-        } else {
-            // Write code directly
-            codeDiffs = await this.editorEngine.sandbox?.getAndWriteCodeDiffs(requests, true);
-        }
-
-        if (codeDiffs.length === 0) {
-            console.error('No code diffs found');
-            return false;
-        }
-
-        return true;
+    async writeRequest(requests: CodeDiffRequest[]) {
+        console.log('writeRequest', requests);
     }
 
     runCodeDiffs(codeDiffs: CodeDiff[]) {
@@ -310,7 +103,9 @@ export class CodeManager {
 
     clear() {
         this.writeQueue = [];
+        this.pendingRequests = [];
         this.isExecuting = false;
+        this.debouncedProcessQueue.cancel();
         this.editorEngine = null as any;
     }
 }
